pyproject.toml
[project]
name = "portfolio_backEnd"
version = "0.1.0"
description = "Default template for PDM package"
authors = [
    {name = "SantiagoSchons", email = "santiagodeoliveira2004@gmail.com"},
]
dependencies = [
    "django>=5.1",
    "djangorestframework>=3.15.2",
    "django-cors-headers>=4.4.0",
    "whitenoise>=6.7.0",
]
requires-python = "==3.10.*"
readme = "README.md"
license = {text = "MIT"}


[tool.pdm]
distribution = false

[tool.pdm.scripts]
runserver = "python manage.py runserver"
makemigrations = "python manage.py makemigrations"
migrate = "python manage.py migrate"